@@ -76,19 +76,19 @@ html_theme = "sphinx_rtd_theme"
 highlight_language = "python"
 
 html_theme_options = {
-    'logo_only': False,
-    'prev_next_buttons_location': 'bottom',
-    'style_external_links': True,
-    'vcs_pageview_mode': 'blob',
-    'style_nav_header_background': '#2980B9',
+    "logo_only": False,
+    "prev_next_buttons_location": "bottom",
+    "style_external_links": True,
+    "vcs_pageview_mode": "blob",
+    "style_nav_header_background": "#2980B9",
     # Toc options
-    'collapse_navigation': True,
-    'sticky_navigation': True,
-    'navigation_depth': 4,
-    'includehidden': True,
-    'titles_only': False,
-    'globaltoc_collapse': True,
-    'globaltoc_maxdepth': 3,
+    "collapse_navigation": True,
+    "sticky_navigation": True,
+    "navigation_depth": 4,
+    "includehidden": True,
+    "titles_only": False,
+    "globaltoc_collapse": True,
+    "globaltoc_maxdepth": 3,
 }
 
 # Add any paths that contain custom static files (such as style sheets) here,
@@ -97,16 +97,16 @@ html_theme_options = {
 html_static_path = ["_static"]
 
 html_css_files = [
-    'custom.css',
+    "custom.css",
 ]
 
 html_show_sourcelink = False
 html_sidebars = {
-    '**': [
-        'globaltoc.html',
-        'relations.html',
-        'sourcelink.html',
-        'searchbox.html',
+    "**": [
+        "globaltoc.html",
+        "relations.html",
+        "sourcelink.html",
+        "searchbox.html",
     ]
 }
 
@@ -130,7 +130,7 @@ autodoc_default_options = {
     "ignore-module-all": True,
     "show-inheritance": True,
     "exclude-members": EXCLUDED_MEMBERS,
-    'inherited-members': False,
+    "inherited-members": False,
 }
 
 # order of members in docs, usefully for methods in class
