@@ -49,12 +49,12 @@ class DataProfiler:
 
     @staticmethod
     def _generate_synthetic_data(
-            n_columns: int,
-            n_rows: int,
-            n2c_ratio: float,
-            rs: int | None,
-            num_range: tuple,
-            n_categories: int,
+        n_columns: int,
+        n_rows: int,
+        n2c_ratio: float,
+        rs: int | None,
+        num_range: tuple,
+        n_categories: int,
     ) -> pd.DataFrame:
         if rs is not None:
             np.random.seed(rs)
@@ -72,11 +72,11 @@ class DataProfiler:
         return pd.DataFrame(
             np.hstack((numerical_data, categorical_data)),
             columns=[f"num_col_{i}" for i in range(n_numerical)]
-                    + [f"cat_col_{i}" for i in range(n_categorical)],
+            + [f"cat_col_{i}" for i in range(n_categorical)],
         )
 
     def create_dataset(
-            self, params: dict
+        self, params: dict
     ) -> tuple[Dataset, dict[str, int | tuple[int, int] | float]]:
         all_params = self.fixed_data_params.copy()
         all_params.update(params)
@@ -92,9 +92,9 @@ class ExperimentProfiler:
     default_experiment_params: ClassVar[dict] = {"n_iterations": 10}
 
     def __init__(
-            self,
-            fixed_experiment_params: dict | None = None,
-            experiment: type = AATest,
+        self,
+        fixed_experiment_params: dict | None = None,
+        experiment: type = AATest,
     ):
         fixed_experiment_params = fixed_experiment_params or {}
         self.fixed_experiment_params = self.default_experiment_params.copy()
@@ -116,12 +116,12 @@ class PerformanceTester:
     resume: ClassVar[defaultdict] = defaultdict(dict)
 
     def __init__(
-            self,
-            dataProfiler: DataProfiler,
-            experimentProfiler: ExperimentProfiler,
-            iterable_params: list | None = None,
-            use_memory: bool = True,
-            rewrite: bool = True,
+        self,
+        dataProfiler: DataProfiler,
+        experimentProfiler: ExperimentProfiler,
+        iterable_params: list | None = None,
+        use_memory: bool = True,
+        rewrite: bool = True,
     ):
         self.dataProfiler = dataProfiler
         self.experimentProfiler = experimentProfiler
@@ -153,14 +153,16 @@ class PerformanceTester:
                     "analysis",
                     *list(self.experimentProfiler.fixed_experiment_params.keys()),
                     *list(self.dataProfiler.fixed_data_params.keys()),
-                    "time", "M1", "M2"
+                    "time",
+                    "M1",
+                    "M2",
                 ]
                 writer.writerow(row_items)
         with alive_bar(
-                self.get_number_params(),
-                bar="squares",
-                spinner="dots_waves2",
-                title=f"Analysis : {analysis}",
+            self.get_number_params(),
+            bar="squares",
+            spinner="dots_waves2",
+            title=f"Analysis : {analysis}",
         ) as bar:
             for params, data, experiment in tqdm(self.get_params()):
                 combined_params = {**data[1], **experiment[1]}
@@ -184,7 +186,7 @@ class PerformanceTester:
                 process.join()
                 monitor.join()
 
-                max_memory_mb = return_dict2["max_memory"] / 1024 ** 2
+                max_memory_mb = return_dict2["max_memory"] / 1024**2
 
                 with open(file_name, "a", newline="") as file:
                     writer = csv.writer(file)
@@ -234,14 +236,14 @@ class PerformanceTester:
 
         return_dict["results"] = [
             exec_time,
-            memory_usage / 10 ** 6 if self.use_memory else None,
+            memory_usage / 10**6 if self.use_memory else None,
         ]
 
 
 def performance_test_plot(
-        params: dict,
-        output_path: str,
-        title="The results of the one-factor performance test of the AA Test",
+    params: dict,
+    output_path: str,
+    title="The results of the one-factor performance test of the AA Test",
 ):
     df = pd.read_csv(output_path)
     df = df[df.analysis == "onefactor"]
