@@ -3,8 +3,11 @@ from __future__ import annotations
 from copy import deepcopy
 from typing import Any, Sequence
 
+import numpy as np
+
 from ..dataset import (
     ABCRole,
+    AdditionalFeatureRole,
     Dataset,
     ExperimentData,
     FeatureRole,
@@ -22,9 +25,11 @@ class MahalanobisDistance(Calculator):
         self,
         grouping_role: ABCRole | None = None,
         key: Any = "",
+        weights: dict[str, float] | None = None,
     ):
         super().__init__(key=key)
         self.grouping_role = grouping_role or GroupingRole()
+        self.weights = weights
 
     @classmethod
     def _execute_inner_function(
@@ -66,7 +71,9 @@ class MahalanobisDistance(Calculator):
 
     def _get_fields(self, data: ExperimentData):
         group_field = data.field_search(self.grouping_role)
-        target_fields = data.field_search(FeatureRole(), search_types=self.search_types)
+        target_fields = data.field_search(
+            [FeatureRole(), AdditionalFeatureRole()], search_types=self.search_types
+        )
         return group_field, target_fields
 
     @property
@@ -74,11 +81,24 @@ class MahalanobisDistance(Calculator):
         return [int, float]
 
     @classmethod
-    def _inner_function(cls, data: Dataset, test_data: Dataset | None = None, **kwargs):
+    def _inner_function(
+        cls,
+        data: Dataset,
+        test_data: Dataset | None = None,
+        weights: dict[str, float] | None = None,
+        **kwargs,
+    ):
         test_data = cls._check_test_data(test_data)
         cov = (data.cov() + test_data.cov()) / 2 if test_data else data.cov()
         cholesky = CholeskyExtension().calc(cov)
         mahalanobis_transform = InverseExtension().calc(cholesky)
+        if weights is not None:
+            features = data.columns
+            w_list = np.array(
+                [weights[col] if col in weights.keys() else 1 for col in features]
+            )
+            w_matrix = np.sqrt(np.diag(w_list / w_list.sum()))
+            mahalanobis_transform = mahalanobis_transform.dot(w_matrix)
         y_control = data.dot(mahalanobis_transform.transpose())
         if test_data:
             y_test = test_data.dot(mahalanobis_transform.transpose())
@@ -92,6 +112,7 @@ class MahalanobisDistance(Calculator):
         group_field: Sequence[str] | str | None = None,
         grouping_data: list[tuple[str, Dataset]] | None = None,
         target_fields: str | list[str] | None = None,
+        weights: dict[str, float] | None = None,
         **kwargs,
     ) -> dict:
         group_field = Adapter.to_list(group_field)
@@ -103,7 +124,11 @@ class MahalanobisDistance(Calculator):
         else:
             raise NotSuitableFieldError(group_field, "Grouping")
         return cls._execute_inner_function(
-            grouping_data, target_fields=target_fields, old_data=data, **kwargs
+            grouping_data,
+            target_fields=target_fields,
+            old_data=data,
+            weights=weights,
+            **kwargs,
         )
 
     def execute(self, data: ExperimentData) -> ExperimentData:
@@ -120,15 +145,17 @@ class MahalanobisDistance(Calculator):
         else:
             grouping_data = None
         t_data = deepcopy(data.ds)
-        if target_fields[1] not in t_data.columns:
-            t_data = t_data.add_column(
-                data.additional_fields[target_fields[1]],
-                role={target_fields[1]: TargetRole()},
-            )
+        for field in target_fields:
+            if field not in t_data.columns:
+                t_data = t_data.add_column(
+                    data.additional_fields[field],
+                    role={field: TargetRole()},
+                )
         compare_result = self.calc(
             data=t_data,
             group_field=group_field,
             target_fields=target_fields,
             grouping_data=grouping_data,
+            weights=self.weights or None,
         )
         return self._set_value(data, compare_result)
