@@ -190,6 +190,10 @@ class PandasNavigation(DatasetBackendNavigation):
             return int
         elif pd.api.types.is_float_dtype(dtype):
             return float
+        elif pd.api.types.is_object_dtype(dtype) and pd.api.types.is_list_like(
+            self.data[column_name].iloc[0]
+        ):
+            return object
         elif (
             pd.api.types.is_string_dtype(dtype)
             or pd.api.types.is_object_dtype(dtype)
@@ -282,6 +286,20 @@ class PandasDataset(PandasNavigation, DatasetBackendCalc):
 
     def __init__(self, data: pd.DataFrame | dict | str | pd.Series | None = None):
         super().__init__(data)
+
+    def get(
+        self,
+        key,
+        default=None,
+    ) -> Any:
+        return self.data.get(key, default)
+
+    def take(
+        self,
+        indices: int | list[int],
+        axis: Literal["index", "columns", "rows"] | int = 0,
+    ) -> Any:
+        return self.data.take(indices=indices, axis=axis)
 
     def get_values(
         self,
@@ -443,8 +461,20 @@ class PandasDataset(PandasNavigation, DatasetBackendCalc):
             return int(data.loc[data.index[0], data.columns[0]])
         return data if isinstance(data, pd.DataFrame) else pd.DataFrame(data)
 
-    def dot(self, other: PandasDataset) -> pd.DataFrame:
-        result = self.data.dot(other.data)
+    def dot(self, other: PandasDataset | np.ndarray) -> pd.DataFrame:
+        if isinstance(other, np.ndarray):
+            other_df = pd.DataFrame(
+                data=other,
+                columns=self.columns if other.shape[1] == self.shape[1] else None,
+            )
+            # print(other_df.shape)
+            # print(self.data.shape)
+            result = self.data.dot(other_df.T)
+            result.columns = (
+                self.columns if other.shape[1] == self.shape[1] else result.columns
+            )
+        else:
+            result = self.data.dot(other.data)
         return result if isinstance(result, pd.DataFrame) else pd.DataFrame(result)
 
     def dropna(
@@ -519,8 +549,13 @@ class PandasDataset(PandasNavigation, DatasetBackendCalc):
             how=how,
         )
 
-    def drop(self, labels: str = "", axis: int = 1) -> pd.DataFrame:
-        return self.data.drop(labels=labels, axis=axis)
+    def drop(
+        self,
+        labels: str | None = None,
+        axis: int | None = None,
+        columns: str | Iterable[str] | None = None,
+    ) -> pd.DataFrame:
+        return self.data.drop(labels=labels, axis=axis, columns=columns)
 
     def filter(
         self,
@@ -547,3 +582,17 @@ class PandasDataset(PandasNavigation, DatasetBackendCalc):
 
     def reindex(self, labels: str = "", fill_value: str | None = None) -> pd.DataFrame:
         return self.data.reindex(labels, fill_value=fill_value)
+
+    def list_to_columns(self, column: str) -> pd.DataFrame:
+        data = self.data
+        n_cols = len(data.loc[0, column])
+
+        data_expanded = (
+            pd.DataFrame(
+                data[column].to_list(), columns=[f"{column}_{i}" for i in range(n_cols)]
+            )
+            if n_cols > 1
+            else data
+        )
+
+        return data_expanded
