@@ -1,3 +1,4 @@
+from .calculators import MinSampleSize
 from .executor import Calculator, Executor, IfExecutor, MLExecutor
 
-__all__ = ["Calculator", "Executor", "IfExecutor", "MLExecutor"]
+__all__ = ["Calculator", "Executor", "IfExecutor", "MLExecutor", "MinSampleSize"]
