@@ -2,10 +2,10 @@ from .encoders import DummyEncoderExtension
 from .faiss import FaissExtension
 from .scipy_linalg import CholeskyExtension, InverseExtension
 from .scipy_stats import (
-                          Chi2TestExtension,
-                          KSTestExtension,
-                          TTestExtension,
-                          UTestExtension,
+    Chi2TestExtension,
+    KSTestExtension,
+    TTestExtension,
+    UTestExtension,
 )
 from .statsmodels import MultiTest, MultitestQuantile
 
