@@ -16,16 +16,19 @@ class FaissExtension(MLExtension):
     ):
         self.n_neighbors = n_neighbors
         self.faiss_mode = faiss_mode
+        self.index = None
         super().__init__()
 
     @staticmethod
     def _prepare_indexes(index: np.ndarray, dist: np.ndarray, k: int):
-        new = [
-            np.concatenate(
-                [val[np.where(dist[i] == d)[0]] for d in sorted(set(dist[i]))[:k]]
-            )
-            for i, val in enumerate(index)
-        ]
+        new = np.vstack(
+            [
+                np.concatenate(
+                    [val[np.where(dist[i] == d)[0]] for d in sorted(set(dist[i]))[:k]]
+                )
+                for i, val in enumerate(index)
+            ]
+        )
         return new
 
     def _predict(self, data: Dataset, test_data: Dataset, X: np.ndarray) -> pd.Series:
@@ -42,7 +45,7 @@ class FaissExtension(MLExtension):
             ]
         else:
             indexes = self._prepare_indexes(indexes, dist, self.n_neighbors)
-        return pd.Series(indexes)
+        return self.result_to_dataset(result=indexes, roles={})
 
     def _calc_pandas(
         self,
@@ -56,10 +59,14 @@ class FaissExtension(MLExtension):
         test = test_data.data.values
         if mode in ["auto", "fit"]:
             self.index = faiss.IndexFlatL2(X.shape[1])
-            if ((
-                len(X) > 1_000_000 and self.faiss_mode == "auto"
-            ) or self.faiss_mode == "fast"
-            ) and len(X) > 1_000 and len(test) > 1_000:
+            if (
+                (
+                    (len(X) > 1_000_000 and self.faiss_mode == "auto")
+                    or self.faiss_mode == "fast"
+                )
+                and len(X) > 1_000
+                and len(test) > 1_000
+            ):
                 self.index = faiss.IndexIVFFlat(self.index, X.shape[1], 1000)
                 self.index.train(X)
             self.index.add(X)
