@@ -1,3 +1,4 @@
+from .cupac import CUPACExecutor
 from .faiss import FaissNearestNeighbors
 
-__all__ = ["FaissNearestNeighbors"]
+__all__ = ["CUPACExecutor", "FaissNearestNeighbors"]
