@@ -4,7 +4,7 @@ from typing import Any, ClassVar
 
 from ..analyzers.ab import ABAnalyzer
 from ..comparators import Chi2Test, TTest, UTest
-from ..dataset import Dataset, ExperimentData
+from ..dataset import Dataset, ExperimentData, StatisticRole
 from ..utils import ExperimentDataEnum
 from .aa import OneAADictReporter
 
@@ -31,7 +31,45 @@ class ABDictReporter(OneAADictReporter):
 class ABDatasetReporter(ABDictReporter):
     @staticmethod
     def _invert_aa_format(table: Dataset) -> Dataset:
-        return table.replace("NOT OK", "N").replace("OK", "NOT OK").replace("N", "OK")
+        return (
+            table
+            if table.is_empty()
+            else table.replace("NOT OK", "N").replace("OK", "NOT OK").replace("N", "OK")
+        )
+
+    def report_variance_reductions(self, data: ExperimentData) -> Dataset | str:
+        """Generate variance reduction report for CUPED/CUPAC transformations."""
+        variance_cols = [
+            col
+            for col in data.additional_fields.columns
+            if col.endswith("_variance_reduction")
+        ]
+        if not variance_cols:
+            return "No variance reduction data available. Ensure CUPED or CUPAC was applied."
+
+        # Create report data
+        report_data = []
+        for col in variance_cols:
+            metric_name = col.replace("_variance_reduction", "")
+            # Get the scalar value from the additional_fields
+            reduction_value = data.additional_fields.data[col].iloc[0]
+            report_data.append(
+                {
+                    "Transformed Metric Name": metric_name,
+                    "Variance Reduction (%)": reduction_value,
+                }
+            )
+
+        # Convert to Dataset
+        if report_data:
+            return Dataset.from_dict(
+                data=report_data,
+                roles={
+                    "Transformed Metric Name": StatisticRole(),
+                    "Variance Reduction (%)": StatisticRole(),
+                },
+            )
+        return "No variance reduction data available."
 
     def report(self, data: ExperimentData):
         front_buffer = self.front
