@@ -25,6 +25,7 @@ class AASplitter(Calculator):
         sample_size: float | None = None,
         constant_key: bool = True,
         save_groups: bool = True,
+        groups_sizes: list[float] | None = None,
         key: Any = "",
     ):
         self.control_size = control_size
@@ -33,6 +34,7 @@ class AASplitter(Calculator):
         self.constant_key = constant_key
         self.save_groups = save_groups
         self.sample_size = sample_size
+        self.groups_sizes = groups_sizes
         super().__init__(key)
 
     def _generate_params_hash(self):
@@ -41,6 +43,8 @@ class AASplitter(Calculator):
             hash_parts.append(f"cs {self.control_size}")
         if self.random_state is not None:
             hash_parts.append(f"rs {self.random_state}")
+        if self.groups_sizes is not None:
+            hash_parts.append(f"gs {self.groups_sizes}")
         self._params_hash = "|".join(hash_parts)
 
     def init_from_hash(self, params_hash: str):
@@ -50,6 +54,12 @@ class AASplitter(Calculator):
                 self.control_size = float(hash_part[hash_part.rfind(" ") + 1 :])
             elif hash_part.startswith("rs"):
                 self.random_state = int(hash_part[hash_part.rfind(" ") + 1 :])
+            elif hash_part.startswith("gs"):
+                self.groups_sizes = []
+                groups_sizes = (
+                    hash_part[hash_part.find(" ") + 1 :].strip("[]").split(",")
+                )
+                self.groups_sizes = [float(gs) for gs in groups_sizes]
         self._generate_id()
 
     @property
@@ -82,6 +92,7 @@ class AASplitter(Calculator):
         data: Dataset,
         random_state: int | None = None,
         control_size: float = 0.5,
+        groups_sizes: list[float] | None = None,
         sample_size: float | None = None,
         const_group_field: str | None = None,
         **kwargs,
@@ -94,9 +105,13 @@ class AASplitter(Calculator):
             if control_data is not None:
                 control_indexes = list(control_data.index)
             const_size = sum(len(cd) for cd in const_data.values())
-            control_size = (len(data) * control_size - const_size) / (
-                len(data) - const_size
+            control_size = (
+                0
+                if len(data) <= const_size
+                else (len(data) * control_size - len(const_data["control"]))
+                / (len(data) - const_size)
             )
+            # control_size = len(data) * control_size
         experiment_data = (
             data[data[const_group_field].isna()] if const_group_field else data
         )
@@ -104,12 +119,34 @@ class AASplitter(Calculator):
             frac=sample_size, random_state=random_state
         ).index
         addition_indexes = list(experiment_data_index)
-        edge = int(len(addition_indexes) * control_size)
-        control_indexes += addition_indexes[:edge]
+        edges = []
+        if groups_sizes:
+            if sum(groups_sizes) != 1:
+                raise ValueError("Groups sizes must sum to 1")
+            for group_size in groups_sizes:
+                size = int(len(addition_indexes) * group_size) + (
+                    0 if not edges else edges[-1]
+                )
+                size = min(size, len(addition_indexes))
+                if size not in edges:
+                    edges += [size]
+        else:
+            edges = [int(len(addition_indexes) * control_size), len(addition_indexes)]
+        control_indexes += addition_indexes[: edges[0]]
+        test_indexes = [
+            addition_indexes[edges[i - 1] : edges[i]] for i in range(1, len(edges))
+        ]
 
-        split_series = pd.Series(np.ones(data.data.shape[0], dtype="int"), index=data.data.index)
+        split_series = pd.Series(
+            np.ones(data.data.shape[0], dtype="int"), index=data.data.index
+        )
         split_series[control_indexes] -= 1
-        split_series = split_series.map({0: "control", 1: "test"})
+        for i, test_index in enumerate(test_indexes):
+            split_series[test_index] += i
+
+        label_map = {0: "control"}
+        label_map.update({i: f"test_{i}" for i in range(1, len(edges))})
+        split_series = split_series.map(label_map)
 
         return split_series.to_list()
 
@@ -124,6 +161,7 @@ class AASplitter(Calculator):
             control_size=self.control_size,
             sample_size=self.sample_size,
             const_group_field=const_group_fields,
+            groups_sizes=self.groups_sizes,
         )
         return self._set_value(
             data,
@@ -144,6 +182,7 @@ class AASplitterWithStratification(AASplitter):
             return AASplitter._inner_function(
                 data, random_state, control_size, **kwargs
             )
+
         result = {"split": []}
         index = []
         for group, group_data in data.groupby(grouping_fields):
@@ -160,6 +199,7 @@ class AASplitterWithStratification(AASplitter):
             random_state=self.random_state,
             control_size=self.control_size,
             grouping_fields=grouping_fields,
+            groups_sizes=self.groups_sizes,
         )
         if isinstance(result, Dataset):
             result = result.replace_roles({"split": AdditionalTreatmentRole()})
