@@ -1,43 +1,50 @@
 from .constants import (
-                        ID_SPLIT_SYMBOL,
-                        MATCHING_INDEXES_SPLITTER_SYMBOL,
-                        NAME_BORDER_SYMBOL,
-                        NUMBER_TYPES_LIST,
+    ID_SPLIT_SYMBOL,
+    MATCHING_INDEXES_SPLITTER_SYMBOL,
+    NAME_BORDER_SYMBOL,
+    NUMBER_TYPES_LIST,
 )
-from .enums import ABNTestMethodsEnum, BackendsEnum, ExperimentDataEnum, SpaceEnum
+from .enums import (
+    ABNTestMethodsEnum,
+    ABTestTypesEnum,
+    BackendsEnum,
+    ExperimentDataEnum,
+    SpaceEnum,
+)
 from .errors import (
-                        AbstractMethodError,
-                        BackendTypeError,
-                        ConcatBackendError,
-                        ConcatDataError,
-                        DataTypeError,
-                        MergeOnError,
-                        NoColumnsError,
-                        NoRequiredArgumentError,
-                        NotFoundInExperimentDataError,
-                        NotSuitableFieldError,
-                        RoleColumnError,
-                        SpaceError,
+    AbstractMethodError,
+    BackendTypeError,
+    ConcatBackendError,
+    ConcatDataError,
+    DataTypeError,
+    MergeOnError,
+    NoColumnsError,
+    NoRequiredArgumentError,
+    NotFoundInExperimentDataError,
+    NotSuitableFieldError,
+    RoleColumnError,
+    SpaceError,
 )
 from .tutorial_data_creation import (
-                        create_test_data,
-                        gen_control_variates_df,
-                        gen_oracle_df,
-                        gen_special_medicine_df,
+    create_test_data,
+    gen_control_variates_df,
+    gen_oracle_df,
+    gen_special_medicine_df,
 )
 from .typings import (
-                        CategoricalTypes,
-                        DecoratedType,
-                        DefaultRoleTypes,
-                        DocstringInheritDecorator,
-                        FromDictTypes,
-                        GroupingDataType,
-                        MultiFieldKeyTypes,
-                        RoleNameType,
-                        ScalarType,
-                        SetParamsDictTypes,
-                        StratificationRoleTypes,
-                        TargetRoleTypes,
+    CategoricalTypes,
+    DecoratedType,
+    DefaultRoleTypes,
+    DocstringInheritDecorator,
+    FeatureRoleTypes,
+    FromDictTypes,
+    GroupingDataType,
+    MultiFieldKeyTypes,
+    RoleNameType,
+    ScalarType,
+    SetParamsDictTypes,
+    StratificationRoleTypes,
+    TargetRoleTypes,
 )
 
 __all__ = [
@@ -46,6 +53,7 @@ __all__ = [
     "NAME_BORDER_SYMBOL",
     "NUMBER_TYPES_LIST",
     "ABNTestMethodsEnum",
+    "ABTestTypesEnum",
     "AbstractMethodError",
     "BackendTypeError",
     "BackendsEnum",
@@ -57,6 +65,7 @@ __all__ = [
     "DefaultRoleTypes",
     "DocstringInheritDecorator",
     "ExperimentDataEnum",
+    "FeatureRoleTypes",
     "FromDictTypes",
     "GroupingDataType",
     "MergeOnError",
